[BUGFIX] use canonical-path of live-server root to allow jetty serving content from a symlinked resource-base
This fix allows serving from MacOS's symlinked temporary directory.
MacOS symlinks `/var` to `/private/var`. The default temporary directory
on MacOS is `/var/folders/…`, so fluentmark-rendered files can not
be served on MacOS, due to jetty's `AllowedResourceAliasChecker`
which prevents serving files from outside the resource-base.

Using the canonical-path (which includes the realpath) of the
live-server's root solves this issue.

Resolves: https://github.com/grosenberg/Fluentmark/issues/73

Signed-off-by: Stephan Jorek <stephan.jorek@gmail.com>